pyproject.toml
[project]
name = "kws"
version = "0.1.0"
description = "基于Python的离线语音关键词检测服务"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sounddevice>=0.4.6",
    "vosk>=0.3.45",
    "pyyaml>=6.0.1",
    "loguru>=0.7.3",
    "numpy>=2.2.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.11.0"
]
